Creates a new 'auth' folder and a new 'auth-router.js' file inside of the 'auth' folder to build a route for the '/register', '/login', '/logout' endpoints
